In mem_manager.hpp, fix the logic for registering the base Memory when creating an alias -- this required GetHypreMemoryLocation() and HypreUsingGPU() to be moved to the same header.
